Refactor InputHandler, refactor handcontroller, refactor animate grasp
The inputhandler now uses the XRDirectInteractor SelectEnter to initiate grasp, i.e the method called when a direct interactor uses the grip button. The HandController has been stripped of the atach transform feature as the XRInteractor does this now. The animate grasp now resets on every animation, now possible to grasp, ungrasp and then grasp again